pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
